Change toc order and create categories
This will add headers in the sidebar and the main page. See docs.readthedocs.io for an example on how this will look.

Created this pull request to discuss the categories. Are all topics in the correct category and order? Are the name of the categories ok?

I'm not sure if application and project is in the correct place. And if sources should be moved to underneath items or not.